Update class name and analysis class name. Move analysis class to existing file.
