+++ qiskit_experiments/library/__init__.py
@@ -77,7 +77,7 @@ Some experiments also have a calibration experiment version.
     ~characterization.FineSXDrag
     ~characterization.MultiStateDiscrimination
     ~characterization.StarkRamseyXY
-    ~characterization.StarkRamseyFast
+    ~characterization.StarkRamseyXYAmpScan
 
 .. _characterization two qubits:
 
@@ -187,7 +187,7 @@ from .characterization import (
     ZZRamsey,
     MultiStateDiscrimination,
     StarkRamseyXY,
-    StarkRamseyFast,
+    StarkRamseyXYAmpScan,
 )
 from .randomized_benchmarking import StandardRB, InterleavedRB
 from .tomography import (
