+++ qiskit_experiments/library/characterization/__init__.py
@@ -50,7 +50,7 @@ Experiments
     MultiStateDiscrimination
     ZZRamsey
     StarkRamseyXY
-    StarkRamseyFast
+    StarkRamseyXYAmpScan
 
 
 Analysis
@@ -69,7 +69,7 @@ Analysis
     DragCalAnalysis
     FineAmplitudeAnalysis
     RamseyXYAnalysis
-    StarkRamseyFastAnalysis
+    StarkRamseyXYAmpScanAnalysis
     ReadoutAngleAnalysis
     ResonatorSpectroscopyAnalysis
     LocalReadoutErrorAnalysis
@@ -83,7 +83,7 @@ from .analysis import (
     DragCalAnalysis,
     FineAmplitudeAnalysis,
     RamseyXYAnalysis,
-    StarkRamseyFastAnalysis,
+    StarkRamseyXYAmpScanAnalysis,
     T2RamseyAnalysis,
     T1Analysis,
     T1KerneledAnalysis,
@@ -108,7 +108,7 @@ from .cr_hamiltonian import CrossResonanceHamiltonian, EchoedCrossResonanceHamil
 from .rabi import Rabi, EFRabi
 from .half_angle import HalfAngle
 from .fine_amplitude import FineAmplitude, FineXAmplitude, FineSXAmplitude, FineZXAmplitude
-from .ramsey_xy import RamseyXY, StarkRamseyXY, StarkRamseyFast
+from .ramsey_xy import RamseyXY, StarkRamseyXY, StarkRamseyXYAmpScan
 from .fine_frequency import FineFrequency
 from .drag import RoughDrag
 from .readout_angle import ReadoutAngle
