+++ releasenotes/notes/mod-stark-1f1afb538a94fe9a.yaml
@@ -5,8 +5,8 @@ features:
     This is a variant of the :class:`.RamseyXY` experiment that characterizes 
     the qubit frequency offset under a Stark tone drive.
   - |
-    New experiment :class:`.StarkRamseyFast` has been added.
+    New experiment :class:`.StarkRamseyXYAmpScan` has been added.
     This is a variant of the :class:`.StarkRamseyXY` experiment to estimate
     the required tone amplitude to cause a particular Stark shift.
     This experiment scans tone amplitude while fixing the Stark tone length,
-    and the fit the result with the dedicated fitter :class:`.StarkRamseyFastAnalysis`.
+    and the fit the result with the dedicated fitter :class:`.StarkRamseyXYAmpScanAnalysis`.
